+++ src/stories/Utilities/Advanced/Cluster/Cluster.js
@@ -1,4 +1,4 @@
-import { createChildren } from '../../helpers/utils'
+import { createChildren } from '../../../helpers/utils'
 
 export const createCluster = ({ cluster = true, alignItems = '', gap = '' }) => {
   const wrapper = document.createElement('div')
