+++ src/stories/Utilities/Advanced/Cluster/Cluster.stories.js
@@ -1,12 +1,12 @@
-import { createStack } from './Stack.js'
+import { createCluster } from './Cluster.js'
 
 export default {
-  title: 'Utilities/Stack',
-  render: ({ stack, ...args }) => {
-    return createStack({ stack, ...args })
+  title: 'Utilities/Advanced/Cluster',
+  render: ({ cluster, ...args }) => {
+    return createCluster({ cluster, ...args })
   },
   argTypes: {
-    stack: { control: 'boolean' },
+    cluster: { control: 'boolean' },
     alignItems: {
       control: { type: 'select' },
       options: ['stretch', 'start', 'center', 'end', 'baseline'],
@@ -23,13 +23,13 @@ export default {
 
 export const With = {
   args: {
-    stack: true,
+    cluster: true,
   },
 }
 
 export const Without = {
   args: {
-    stack: false,
+    cluster: false,
   },
 }
