+++ src/stories/Utilities/Advanced/Stack/Stack.js
@@ -1,4 +1,4 @@
-import { createChildren } from '../../helpers/utils'
+import { createChildren } from '../../../helpers/utils'
 
 export const createStack = ({ stack = true, alignItems = '', gap = '' }) => {
   const wrapper = document.createElement('div')
