+++ src/stories/Utilities/Advanced/Stack/Stack.mdx
@@ -1,10 +1,10 @@
 import { Meta, Story, Canvas, Controls } from '@storybook/addon-docs/blocks'
-import * as ClusterStories from './Cluster.stories'
-import { createSourceCodeLink } from '../../helpers/sourceCodeLink.js'
+import * as StackStories from './Stack.stories'
+import { createSourceCodeLink } from '../../../helpers/sourceCodeLink.js'
 
-<Meta of={ClusterStories} />
+<Meta of={StackStories} />
 
-# Cluster
+# Stack
 
 <div
   dangerouslySetInnerHTML={{
@@ -12,11 +12,11 @@ import { createSourceCodeLink } from '../../helpers/sourceCodeLink.js'
   }}
 ></div>
 
-The cluster utility provides a simple way to create a wrapping, group of spaced items. It
+The stack utility provides a simple way to create a vertical stack of spaced items. It
 works with all of the gap and other flex utilities. It is effectively equivalent to using
-`.flex .flex-wrap .items-center .gap-md` together. See [Utility Introduction](?path=/docs/utilities-introduction--docs#higher-order-utilities-vs-components) for more information.
+`.flex .flex-col .gap-md` together. See [Utility Introduction](?path=/docs/utilities-introduction--docs#higher-order-utilities-vs-components) for more information.
 
-See [Utility Layout](?path=/docs/recipes-layout--docs#utility) for an example of how clusters
+See [Utility Layout](?path=/docs/recipes-layout--docs#utility) for an example of how stacks
 can be used to create more readable flex layouts.
 
 Note: This utility uses the `op` prefix to avoid potential naming conflicts with other CSS frameworks.
@@ -24,47 +24,47 @@ This is a pattern we hope to move towards for all utilities in the future.
 
 ## Playground
 
-<Canvas of={ClusterStories.With} />
-<Controls of={ClusterStories.With} />
+<Canvas of={StackStories.With} />
+<Controls of={StackStories.With} />
 
 ## Without
 
-A normal `div` without the cluster utility
+A normal `div` without the stack utility
 
-<Canvas of={ClusterStories.Without} />
+<Canvas of={StackStories.Without} />
 
-## Cluster Property
+## Stack Property
 
-`.op-cluster` Creates a flex cluster.
+`.op-stack` Creates a flex stack.
 
-<Canvas of={ClusterStories.With} />
+<Canvas of={StackStories.With} />
 
 ## Align Items Stretch
 
 `.items-stretch` Stretches each item across the cross axis. This is the default alignment and doesn't need to be applied unless you are overriding something set differently.
 
-<Canvas of={ClusterStories.AlignStretch} />
+<Canvas of={StackStories.AlignStretch} />
 
 ## Align Items Start
 
 `.items-start` places each item at the start of the cross axis.
 
-<Canvas of={ClusterStories.AlignStart} />
+<Canvas of={StackStories.AlignStart} />
 
 ## Align Items Center
 
 `.items-center` places each item at the center of the cross axis.
 
-<Canvas of={ClusterStories.AlignCenter} />
+<Canvas of={StackStories.AlignCenter} />
 
 ## Align Items End
 
 `.items-end` places each item at the end of the cross axis.
 
-<Canvas of={ClusterStories.AlignEnd} />
+<Canvas of={StackStories.AlignEnd} />
 
 ## Align Items Baseline
 
 `.items-baseline` places each item at the text baseline of the cross axis.
 
-<Canvas of={ClusterStories.AlignBaseline} />
+<Canvas of={StackStories.AlignBaseline} />
