--- src/stories/Utilities/Split/Split.mdx
@@ -1,70 +1,0 @@
-import { Meta, Story, Canvas, Controls } from '@storybook/addon-docs/blocks'
-import * as SplitStories from './Split.stories'
-import { createSourceCodeLink } from '../../helpers/sourceCodeLink.js'
-
-<Meta of={SplitStories} />
-
-# Split
-
-<div
-  dangerouslySetInnerHTML={{
-    __html: createSourceCodeLink({ link: 'core/utilities.css' }).outerHTML,
-  }}
-></div>
-
-The split utility provides a simple way to push two items apart from each other.
-It is effectively equivalent to using `.flex .flex-wrap .items-center .justify-between .gap-md` together.
-See [Utility Introduction](?path=/docs/utilities-introduction--docs#higher-order-utilities-vs-components) for more information.
-
-See [Utility Layout](?path=/docs/recipes-layout--docs#utility) for an example of how splits
-can be used to create more readable flex layouts.
-
-Note: This utility uses the `op` prefix to avoid potential naming conflicts with other CSS frameworks.
-This is a pattern we hope to move towards for all utilities in the future.
-
-## Playground
-
-<Canvas of={SplitStories.With} />
-<Controls of={SplitStories.With} />
-
-## Without
-
-A normal `div` without the split utility
-
-<Canvas of={SplitStories.Without} />
-
-## Split Property
-
-`.op-split` Creates a flex split.
-
-<Canvas of={SplitStories.With} />
-
-## Align Items Stretch
-
-`.items-stretch` Stretches each item across the cross axis. This is the default alignment and doesn't need to be applied unless you are overriding something set differently.
-
-<Canvas of={SplitStories.AlignStretch} />
-
-## Align Items Start
-
-`.items-start` places each item at the start of the cross axis.
-
-<Canvas of={SplitStories.AlignStart} />
-
-## Align Items Center
-
-`.items-center` places each item at the center of the cross axis.
-
-<Canvas of={SplitStories.AlignCenter} />
-
-## Align Items End
-
-`.items-end` places each item at the end of the cross axis.
-
-<Canvas of={SplitStories.AlignEnd} />
-
-## Align Items Baseline
-
-`.items-baseline` places each item at the text baseline of the cross axis.
-
-<Canvas of={SplitStories.AlignBaseline} />
